Fixed errors

+++ rejuvenation/setup/setupVm.sh
@@ -1,4 +1,10 @@
 #!/bin/bash
+echo "Have you copied the VM vmDebian.ova file? It should have been put one level before the rejuvenation folder(y/n)"
+read -r copy
+
+if [ "$copy" != "y" ]; then
+  exit 1
+fi
 
 vboxmanage controlvm vmDebian poweroff
 VBoxManage unregistervm vmDebian --delete
@@ -20,13 +26,15 @@ VBoxManage modifyvm vmDebian --natpf1 "porta 8080,tcp,$HOST_IP,8080,,80"
 
 vboxmanage startvm vmDebian --type headless
 
-sleep 20
+sleep 30
 
 ssh-copy-id -i /root/.ssh/id_rsa.pub -p 2222 root@localhost
 
 curl http://localhost:8080
 
+echo "Would you like to start the test? (y/n)"
+read -r start
 
-
-
-
+if [ "$start" = "y" ]; then
+  ./rejuvenation/run.sh -n vmDebian -ssh
+fi
